+++ run.py
@@ -1,0 +1,4 @@
+from bot import create_bot
+
+if __name__ == "__main__":
+    create_bot().run_polling()
